+++ PDCurses++/Source/ConsoleStyle.cpp
@@ -3,18 +3,6 @@
 
 
 
-//ConsoleStyle::ConsoleStyle(std::string name, short id, Color fg, Color bg)
-//{
-//	this->Background = bg;
-//	this->Foreground = fg;
-//	this->ID = id;
-//	this->name = name;
-//	this->Attributes = 0;
-//	
-//
-//	init_pair(id, (short)fg, (short)bg);
-//}
-
 ConsoleStyle::ConsoleStyle(std::string name, short id, Color fg, Color bg, std::initializer_list<Attribute> attrs)
 {
 	this->name = name;
@@ -24,7 +12,7 @@ ConsoleStyle::ConsoleStyle(std::string name, short id, Color fg, Color bg, std::
 	init_pair(id, (short)fg, (short)bg);
 
 	for (auto att : attrs)
-		Attributes |= (int32_t)att;
+		Attributes |= (unsigned long)att;
 }
 
 
@@ -52,7 +40,7 @@ std::string ConsoleStyle::GetName()
 	return this->name;
 }
 
-int32_t ConsoleStyle::GetAttributes()
+unsigned long ConsoleStyle::GetAttributes()
 {
 	return this->Attributes;
 }
